requirements.txt
docker==3.7.3
docker-compose==1.24.1
docker-pycreds==0.4.0
dockerpty==0.4.1
django==2.2.4
environment-kernels==1.1.1
imageio==2.3.0
imagesize==1.0.0
imutils==0.5.2
ipykernel==4.8.2
ipyparallel==6.2.2
ipython==6.4.0
ipython-genutils==0.2.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.32.1
jupyterlab-launcher==0.10.5
matplotlib==3.0.3
msgpack==0.6.0
msgpack-numpy==0.4.3.2
msgpack-python==0.5.6
nltk==3.3
notebook==5.5.0
numpy==1.15.4
opencv-python==3.4.2.17
packaging==17.1
Pillow==5.2.0
pytest==3.5.1
pytest-arraydiff==0.2
pytest-astropy==0.3.0
pytest-doctestplus==0.1.3
pytest-openfiles==0.3.0
pytest-remotedata==0.2.1
python-dateutil==2.7.3
requests==2.20.0
scikit-image==0.14.2
scikit-learn==0.20.3
testpath==0.3.1
textblob==0.15.3
texttable==0.9.1
torchvision==0.2.1
urllib3==1.23
lmdb==0.96
gunicorn==19.9.0
natsort==6.0.0
config==0.4.2
scikit-learn==0.20.3
scikit-image==0.14.2
pandas==0.24.2
lmdb==0.96
pillow==5.2.0
torchvision==0.2.1
natsort==6.0.0
